Fix remaining tests.
