+++ python/cudf/cudf/tests/test_parquet.py
@@ -28,7 +28,6 @@ from cudf.testing._utils import (
     TIMEDELTA_TYPES,
     assert_eq,
     assert_exceptions_equal,
-    random_bitmask,
 )
 
 
@@ -2124,7 +2123,7 @@ def test_parquet_writer_statistics(tmpdir, pdf, add_nulls):
     gdf = cudf.from_pandas(pdf)
     if add_nulls:
         for col in gdf:
-            gdf[col] = gdf[col].set_mask(random_bitmask(len(gdf)))
+            gdf[col][np.random.choice([False, True], size=len(gdf))] = None
     gdf.to_parquet(file_path, index=False)
 
     # Read back from pyarrow
